requirements.txt
streamlit==1.46.1
openai==1.93.0
sentence-transformers==5.0.0
faiss-cpu==1.11.0
python-dotenv==1.1.1
numpy==2.3.1
huggingface-hub==0.33.1
torch==2.7.1
transformers==4.53.0